Fixed MM5 mapper bug that was causing corruption of BG tiles in Castlevania 3.. Slight optimization of the render buffer to 16-bit to reduce cache penalties, and rendering logic for MMC5 mappers.

+++ src/cores/virtuanes/3ds/palette.h
@@ -12,7 +12,7 @@ typedef struct
 } SNesPalette;
 
 extern SNesPalette     nesPalette[64];
-extern u32             rgbaPalette[16][64];
+extern u16             rgbaPalette[16][64];
 
 void nespalInitialize();
